Use separator and remove special status of hyphen
Though the separator can be overridden, this has been using the default
separator ('-') behind the scenes then replacing it with the chosen
separator as the last step. This means, among other things, that hyphens
would always be replaced with the separator, even if you overrode the
allowed characters regex to try to allow them.

This changes `slugify` to use the provided `separator` throughout, which
required some changes to the allowed-characters and multiple-separators
logic.